wait will cause the process to dissapear so only kill if we have not waited

+++ fastprocess/fastprocess.py
@@ -3,6 +3,7 @@ import signal
 
 class FastProcess:
     def __init__(self, cmd, stdin=None, stdout=None, stderr=None, env=None):
+        self.waited = False
         file_actions = []
         if stdin:
             file_actions.append((os.POSIX_SPAWN_DUP2, stdin.fileno(), 0))
@@ -13,7 +14,8 @@ class FastProcess:
         self.pid = os.posix_spawnp(cmd[0], cmd, env if env else os.environ, file_actions=file_actions)
 
     def __del__(self):
-        os.kill(self.pid, signal.SIGTERM)
+        if not self.waited:
+            os.kill(self.pid, signal.SIGTERM)
 
     def terminate(self):
         os.kill(self.pid, signal.SIGTERM)
@@ -22,6 +24,7 @@ class FastProcess:
         os.kill(self.pid, sig)
 
     def wait(self):
+        self.waited = True
         return os.WEXITSTATUS(os.waitpid(self.pid, 0)[1])
